requirements.txt

flask
python-dotenv
google-generativeai
chromadb
pypdf
requests
pyyaml
